+++ packages/grpc-mock/script/grpc-mock-server-start.js
@@ -1,4 +1,0 @@
-const { commandServerStartCli } = require('../lib/index.cjs');
-const { version } = require('../package.json');
-
-commandServerStartCli(version);