+++ packages/mock/script/http.js
@@ -1,4 +1,0 @@
-const { commandHttpCli } = require('../lib/index.cjs');
-const { version } = require('../package.json');
-
-commandHttpCli(version);